Only reserve 8 bits for storing area attributes flags.

+++ libosmscout/src/osmscout/Area.cpp
@@ -106,7 +106,7 @@ namespace osmscout {
 
   bool AreaAttributes::Read(FileScanner& scanner)
   {
-    uint16_t flags;
+    uint8_t flags;
 
     scanner.ReadNumber(type);
     scanner.Read(flags);
